active ticket model seperated from the widgets
