+++ lib/model/widgets.dart
@@ -103,134 +103,7 @@ class HorizontalLineWithOr extends StatelessWidget {
   }
 }
 
-//Active Ticket
-
-class activeTicket extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Container(
-        decoration: const BoxDecoration(
-          color: Colors.white,
-          boxShadow: [
-            BoxShadow(
-              color: Colors.grey,
-              blurRadius: 5.0,
-            ),
-          ],
-        ),
-        width: 330,
-        child: Column(
-          children: [
-            Expanded(
-              flex: 4,
-              child: Container(
-                padding: EdgeInsets.all(10),
-                height: double.infinity,
-                width: double.infinity,
-                child: Row(
-                  children: [
-                    Expanded(
-                      flex: 2,
-                      child: SizedBox(
-                        height: double.infinity,
-                        width: double.infinity,
-                        child: Row(
-                          children: [
-                            Expanded(
-                              flex: 3,
-                              child: SizedBox(
-                                height: double.infinity,
-                                width: double.infinity,
-                                child: Column(
-                                  mainAxisAlignment:
-                                      MainAxisAlignment.spaceAround,
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: [
-                                    Text(
-                                      'SOE',
-                                      style: kSOEStyle,
-                                    ),
-                                    Image.asset('./assets/doubleArrow.png'),
-                                    Text(
-                                      'Pathadipalam',
-                                      style: kDestinationStyle,
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                            Expanded(
-                              flex: 2,
-                              child: SizedBox(
-                                height: double.infinity,
-                                width: double.infinity,
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  mainAxisAlignment:
-                                      MainAxisAlignment.spaceAround,
-                                  children: [
-                                    Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
-                                        Text('Issue date', style: kSmallGrey),
-                                        Text(
-                                          '10 jan,23',
-                                          style: kLargeGrey,
-                                        ),
-                                      ],
-                                    ),
-                                    Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
-                                        Text('Route', style: kSmallGrey),
-                                        Text('Vytilla', style: kLargeGrey),
-                                      ],
-                                    ),
-                                    Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
-                                        Text('Expiry date', style: kSmallGrey),
-                                        Text(
-                                          '9 Feb, 23',
-                                          style: TextStyle(
-                                            fontWeight: FontWeight.w500,
-                                            color: kGreyTextColor,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                    Expanded(
-                      child: Image(
-                        image: AssetImage('./assets/qrcode.png'),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            Text(
-              '#Ticket ID',
-              style: TextStyle(fontWeight: FontWeight.w500, fontSize: 15),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
+//Custom widget for home page dot indicator for the number of active tickets
 
 class DotIndicator extends StatelessWidget {
   final int itemCount;
@@ -242,7 +115,7 @@ class DotIndicator extends StatelessWidget {
     Key? key,
     required this.itemCount,
     required this.controller,
-    this.activeColor = Colors.green,
+    this.activeColor = kGreenMainTheme,
     this.inactiveColor = Colors.grey,
   }) : super(key: key);
 
@@ -269,21 +142,3 @@ class DotIndicator extends StatelessWidget {
     );
   }
 }
-
-//activeticket list
-// class ActiveTicketList extends StatelessWidget {
-//   final List<activeTicket> activeTickets;
-//
-//   const ActiveTicketList({Key? key, required this.activeTickets})
-//       : super(key: key);
-//
-//   @override
-//   Widget build(BuildContext context) {
-//     return ListView.builder(
-//       itemCount: activeTickets.length,
-//       itemBuilder: (context, index) {
-//         return activeTickets[index];
-//       },
-//     );
-//   }
-// }
